test: move legacy tests to non-depecated class tests
